+++ backend/src/DataFixtures/SystemVariablesFixtureLoader.php
@@ -58,6 +58,7 @@ class SystemVariablesFixtureLoader
                 $imagePrefix . base64_encode((string) file_get_contents($pathToImages . '/logo_300.png')),
                 SystemVariable::SCOPE_PUBLIC
             ],
+            SystemVariable::CUSTOMIZATION_LOGIN_TEXT            => ['', SystemVariable::SCOPE_PUBLIC],
             SystemVariable::CUSTOMIZATION_HOME_MARKDOWN         => ['', SystemVariable::SCOPE_PUBLIC],
             SystemVariable::CUSTOMIZATION_FOOTER_TEXT           => [
                 'Documentation is available on GitHub.',